backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.1
langchain==0.3.19
langchain-core==0.3.40
langchain-ollama==0.2.3
ollama==0.4.7
pandas==2.2.3
pdfminer.six==20231228
pdfminer.six==20191110
pdfplumber==0.11.5
python-docx==1.1.2
textract==1.6.5